4.22 tehty

+++ osa4/controllers/blogs.js
@@ -1,6 +1,6 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
-const User = require('../models/user')
+const userExtractor = require('../utils/middleware').userExtractor
 
 blogsRouter.get('/', async (request, response) => {
     const blogs = await Blog
@@ -9,11 +9,13 @@ blogsRouter.get('/', async (request, response) => {
     response.json(blogs)
 })
 
-blogsRouter.post('/', async (request, response) => {
+blogsRouter.post('/', userExtractor, async (request, response) => {
     const body = request.body
+    const user = request.user
 
-    const users = await User.find({})
-    const user = users[0]
+    if (!user) {
+        return response.status(401).json({ error: 'invalid token' })
+    }
 
     const blog = new Blog({
         title: body.title,
@@ -39,10 +41,21 @@ blogsRouter.put('/:id', async (request, response) => {
     }
 
     const result = await Blog.findByIdAndUpdate(request.params.id, blog, { new: true })
-    response.json(result)
+    response.json(result).end()
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
+blogsRouter.delete('/:id', userExtractor, async (request, response) => {
+    const user = request.user
+    const blog = await Blog.findById(request.params.id)
+
+    if (!blog) {
+        return response.status(404).json({ error: 'Blog not found' })
+    }
+
+    if (blog.author.toString() !== user.id.toString()) {
+        return response.status(401).json({ error: 'unauthorized' })
+    }
+
     await Blog.findByIdAndDelete(request.params.id)
     response.status(204).end()
 })
